[REFACTOR] 배당금 배치작업 4개를 2개로 통합
