+++ src/main/java/com/animalfarm/mlf/config/batch/DividendClosingJobConfig.java
@@ -27,7 +27,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Configuration
 @RequiredArgsConstructor
-public class DividendClosingBatchConfig {
+public class DividendClosingJobConfig {
 
 	private final JobBuilderFactory jobBuilderFactory;
 	private final StepBuilderFactory stepBuilderFactory;
